Added new settings system and debug mode

+++ sheets.py
@@ -7,9 +7,9 @@ The sheet in question should be the current year's IEEE Membership sheet.
 '''
 
 from datetime import datetime
-import tomllib
 import gspread
 from log import logger
+from settings import settings
 
 SCOPES = ['https://www.googleapis.com/auth/drive.file']
 
@@ -17,9 +17,7 @@ def get_spreadsheet_id() -> str:
     '''
     Gets the spreadsheet ID from the auth.toml file.
     '''
-    with open('auth.toml', 'rb') as f:
-        auth = tomllib.load(f)
-        return auth['GoogleSheets']['spreadsheet_id']
+    return settings['GoogleSheets']['spreadsheet_id']
 
 def get_worksheet() -> gspread.Worksheet:
     '''
@@ -45,14 +43,15 @@ def get_list_of_known_members() -> list[dict[str, str]]:
     
     # create a list of dictionaries, where each dictionary contains the attributes 'name', 'email', 'status', and 'status_date'
     member_info = []
-    for names, emails, status, status_dates in zip(names_list, emails_list, status_list, status_dates_list):
-        member_info.append({
-            'name': names,
-            'email': emails,
-            'status': status,
-            'status_date': status_dates
-        })
-    logger.debug(f'Found known members.')
+    if settings.get('Debug') != True:
+        for names, emails, status, status_dates in zip(names_list, emails_list, status_list, status_dates_list):
+            member_info.append({
+                'name': names,
+                'email': emails,
+                'status': status,
+                'status_date': status_dates
+            })
+        logger.debug(f'Found known members.')
     return member_info
 
 def add_prospective_member_to_sheet(member: dict[str, str]):
@@ -69,7 +68,8 @@ def add_prospective_member_to_sheet(member: dict[str, str]):
     current_date = datetime.now().strftime('%m/%d/%y')
 
     # add the new member to the sheet
-    worksheet.append_row([member['name'], member['email'], '', 'EMAIL SENT', current_date], table_range='A1:E1')
+    if settings.get('Debug') != True:
+        worksheet.append_row([member['name'], member['email'], '', 'EMAIL SENT', current_date], table_range='A1:E1')
 
     logger.debug(f'Added new member {member["name"]} to sheet.')
 
@@ -88,8 +88,9 @@ def update_member_status(member: dict[str, str], new_status: str):
     # update the status of the member in the sheet
     cell = worksheet.find(member['email'])
     member_row = cell.row
-    worksheet.update_cell(member_row, 4, new_status)
-    worksheet.update_cell(member_row, 5, current_date)
+    if settings.get('Debug') != True:
+        worksheet.update_cell(member_row, 4, new_status)
+        worksheet.update_cell(member_row, 5, current_date)
 
     logger.debug(f'Updated member status of {member["name"]} to {new_status} in the sheet.')
 
@@ -108,9 +109,10 @@ def member_approved(member: dict[str, str], member_id: str):
     # update the status of the member in the sheet
     cell = worksheet.find(member['email'])
     member_row = cell.row
-    worksheet.update_cell(member_row, 3, member_id)
-    worksheet.update_cell(member_row, 4, 'APPROVED')
-    worksheet.update_cell(member_row, 5, current_date)
+    if settings.get('Debug') != True:
+        worksheet.update_cell(member_row, 3, member_id)
+        worksheet.update_cell(member_row, 4, 'APPROVED')
+        worksheet.update_cell(member_row, 5, current_date)
 
     logger.debug(f'Updated member status of {member["name"]} to APPROVED in the sheet with id {member_id}.')
 
@@ -126,7 +128,8 @@ def remove_member(member: dict[str, str]):
     # update the status of the member in the sheet
     cell = worksheet.find(member['email'])
     member_row = cell.row
-    worksheet.delete_rows(member_row, member_row)
+    if settings.get('Debug') != True:
+        worksheet.delete_rows(member_row, member_row)
 
     logger.debug(f'Removed member {member["name"]} from sheet.')
